- Удален legacy класс PersonalData, а так же изменено имя файла с классом Education
